// frontend/src/api/generated/model/postListItem.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface PostListItem {
  id?: number;
  accountId?: number;
  platform?: string;
  externalPostId?: string;
  postType?: string;
  permalink?: string;
  caption?: string;
  remoteThumbnailUrl?: string;
  publishedAt?: string;
  sortValue?: number;
}